Implemented parent MediaSession at SFU mcs-core and did some cleanup
